07_Book_Recommendation_Agent/pyproject.toml
[project]
name = "07-book-recommendation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
    "streamlit>=1.46.1",
]
